+++ packages/studio-be/src/studio/nlu/bot-factory.ts
@@ -1,13 +1,12 @@
 import { Logger } from 'botpress/sdk'
 import _ from 'lodash'
-import { DefinitionsRepository } from '../definitions-repository'
-import { ModelEntryService, TrainingEntryService, ModelEntryRepository } from '../model-entry'
-import { NLUClient } from '../nlu-client'
-import pickSeed from '../pick-seed'
+import { Bot } from './bot'
+import { DefinitionsRepository } from './definitions-repository'
+import { ModelEntryService, TrainingEntryService, ModelEntryRepository } from './model-entry'
+import { NLUClient } from './nlu-client'
+import pickSeed from './pick-seed'
 
-import { BotDefinition, BotConfig, ConfigResolver, TrainListener } from '../typings'
-import { BotState } from './bot-state'
-import { BotStateMachine } from './bot-state-machine'
+import { BotDefinition, BotConfig, ConfigResolver, TrainListener } from './typings'
 
 const CLOUD_NLU_ENDPOINT = process.env.CLOUD_NLU_ENDPOINT || 'https://nlu.botpress.dev'
 
@@ -21,7 +20,7 @@ export class BotFactory {
     private _nluEndpoint: string
   ) {}
 
-  public makeBot = async (botConfig: BotConfig): Promise<BotStateMachine> => {
+  public makeBot = async (botConfig: BotConfig): Promise<Bot> => {
     const { id: botId, cloud } = botConfig
 
     const baseURL = cloud ? CLOUD_NLU_ENDPOINT : this._nluEndpoint
@@ -44,14 +43,14 @@ export class BotFactory {
 
     const modelService = new ModelEntryService(this._modelStateRepo)
     const trainService = new TrainingEntryService(this._modelStateRepo)
-    const botState = new BotState(
+    return new Bot(
       botDefinition,
       this._configResolver,
       nluClient,
       this._defRepo,
       modelService,
-      trainService
+      trainService,
+      this._webSocket
     )
-    return new BotStateMachine(botState, this._defRepo, this._webSocket)
   }
 }
